kaloriju kalkulators salabots, izmaiņas spGetCaloriesByName)

+++ src/main/java/lv/dita/project/data/FoodEaten.java
@@ -16,7 +16,7 @@ public class FoodEaten implements EntityBase {
     private int id;
     private String name;
     private double quantity;
-    private int calories;
+    private double calories;
 
     public static final String SELECT_QUERY = "select * from " + SCHEMA_NAME + ".food_eaten";
 
@@ -26,7 +26,7 @@ public class FoodEaten implements EntityBase {
             foodEaten = new FoodEaten(rs.getInt("food_eaten_id"),
                     rs.getString("food_eaten_name"),
                     rs.getDouble("food_eaten_quantity"),
-                    rs.getInt("food_eaten_calories"));
+                    rs.getDouble("food_eaten_calories"));
 
         } catch (SQLException e) {
             e.printStackTrace();
